refactor : 리뷰반영(프레임 점수 계산 책임 변경)
