+++ src/main/java/bowling/model/NormalFrame.java
@@ -4,6 +4,7 @@ import bowling.model.framestatus.FrameStatus;
 import bowling.model.framestatus.RequiredFirstRoll;
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.IntStream;
 
 public class NormalFrame implements Frame {
 
@@ -60,6 +61,29 @@ public class NormalFrame implements Frame {
   @Override
   public FrameStatus getFrameStatus() {
     return frameStatus;
+  }
+
+  @Override
+  public int getScoreBy(List<Frame> frames) {
+    int startIndex = getFirstIndexOfScoredFrames();
+    int lastIndex = getLastIndexOfScoredFrames();
+
+    if (frames.size() < lastIndex) {
+      lastIndex = frames.size();
+    }
+
+    return IntStream.range(startIndex, lastIndex)
+        .map(index -> frames.get(index).getPins().getFirstKnockDownNumber())
+        .sum() + frames.get(startIndex).getPins().getSecondKnockDownNumber();
+  }
+
+  private int getFirstIndexOfScoredFrames() {
+    return frameStatus.getScoringFramesIndexes().get(0);
+  }
+
+  private int getLastIndexOfScoredFrames() {
+    return frameStatus.getScoringFramesIndexes().get(0) + frameStatus.getScoringFramesIndexes()
+        .size();
   }
 
   @Override
